Clean up dependabot utility
Lock dependency to latest version of go.opentelemetry.io/build-tools to
ensure the FindMods function is there. The previous release built
against a dependency that did not include it.

Fix the label for Go updates to be `go` not `actions`.

Use and indent of two spaces to match existing configs.