[dbotconf] Update all detected Dockerfile
Currently the dbotconf command only adds an entry to update the
Dockerfile at the root of a repository. There are projects, like the
opentelemetry-go repository, that have many Dockerfiles that exist in
subdirectories to the root.

This updates the command to discover all Dockerfiles and add an entry
for them when generating and verify an entry with verifying.