[multimod] add support for --tag to sync
This allows end users to pass in an optional argument when running `multimod sync` if they want to fetch a version that matches a specific tag. This will be useful in the collector contrib for example to sync the core modules, without the user having to manually set the versions in versions.yaml.

Signed-off-by: Alex Boten <aboten@lightstep.com>